requirements.txt
pyspark
numpy
pandas==2.2.3
matplotlib
python-dotenv
certifi
pyarrow
pymongo
requests
PyYAML
ipykernel
scikit-learn

azure-mgmt-compute
azure-mgmt-storage
azure-mgmt-resource
azure-keyvault-secrets
azure-storage-blob
azureml
azureml-core
